Handle the query string being empty
We were returning an error becuase we expected the Rummager response to be a
hash, but it was an array. This is because gds-api-adapters was "helpfully"
returning the equivalent of 0 results if the query was blank.

design-principles has been broken in this respect since:
19d4ec72279f73138a77e092d278d49dbda6803f

Although it doesn't explicitly tell the user why they've got no results, it
should make sense that blank query equals no results.