Users don't recognise words from shapes.
In fact the opposite is true, according to the very document used to support this statement. It says: "Word shape is no longer a viable model of word recognition."

*Detailed Explanation*

The link goes to http://www.microsoft.com/typography/ctfonts/WordRecognition.aspx 
The document starts with an informal introduction in which the author writes:

“There I learned to my chagrin that we recognize words from their word shape and that “Modern psychologists call this image the ‘Bouma shape.’” "

However, in the main paper the writer decribes a wide sweep of evidence and then concludes:

"Word shape is no longer a viable model of word recognition. The bulk of scientific evidence says that we recognize a word’s component letters, then use that visual information to recognize a word.”