Remove display_content_formats? feature flag
We have had this rolled out everywhere for a long time, and it’s stable.

The next major change expected (hoped for!) is to get this into a
published manual format in the normal publishing tools. **When** that
happens, we can archive part of this application to the attic. \o/