Transformation dashboard version-48
Further detail in pull request:
https://github.com/alphagov/transformation-dashboard/pull/82

This makes many changes to the Transformation Dashboard to mark the end 
of the Transformation Programme.

We've made one final round of updates to the exemplar pages, and then 
'archived' them. This means that they now live under the 
`/transformation/exemplars/` path, and are marked with a blue banner 
at the top of the page which tells users that they are no longer being 
updated.

We've also archived the original dashboard page, moving it to 
`exemplars/index`. In its place, we have a new item of content which 
gives a reflection on the programme, and some links to the live services.