Don't catch one error and trigger another
Currently, if the call to rummager fails for any reason, the error is
caught, and @results is set to `["HELP!"]`.  This then fails in the
view, with `undefined method `link' for "HELP!"`.

Instead of trying to present a "fake" result list containing an error,
which will be unhelpful to users, and also get cached for a while, just
let the original error bubble up.  This will cause the app to give a 5xx
error page, which correctly tells the user to try again.  It might also
be caught by the cache in some cases and not even be shown to the user.