Register special routes and redirects via publishing API
we've introduced a new special_route format[1] to allow us to register
unusual routes which don't have a real format associated with them.

This adds a rake task to register the special routes via this mechanism.

We also register all redirects via the publishing-api as well.

We have retained an alias for the `rake router:register` task to this new task
for backward compatability to ease the deployment of this change.

[1] alphagov/govuk-content-schemas#106