Upgrade Slimmer, explicitly set 'wrapper' layout
This should be a no-op change to app behaviour. The Slimmer default was 'wrapper'
before the upgrade, and is not explicitly set.

Slimmer 9.0.0 switched to `core_layout` [1] as the default layout, which is the
modern/recommend layout. Apps not yet using `core_layout` should be explicitly
marked as such, so we can easily identify them as needing updating.

I had a quick look at what would be required to use a modern layout and it's not
immediately clear what static styles this app relies upon.

I think it might be safe to switch to `core_layout`, and that the layout doesn't
really matter to thsi app, as so many parts are complete HTML pages, or have
such heavy custom styling that it doesn't matter?