Hide cookie/survey bars for wraith comparisons
These appear intermitently and were causing noise in the diffs, so
i've marked them both as seen.

I've tried to keep the cookie domain general, rather than app specific
as we'll likely use this as a template for other apps using wraith,
though it will only work for VM URLs.