Use string paths instead of hash options for nav_link_to
The rails 4 router is more strict about trying to build routes that aren't
defined in routes.rb, and the pages that render the nav_link_to helpers
were trying to render routes that didn't exist.  The current_page? helper
can take a string instead of a url hash, so we change to use that and this
means we're not trying to generate routes that don't exist.

This will probably also fix the issue with the live page where the urls in
the rendered links are of the form:

  https://www.gov.uk/design-principles/mainstream?action=mainstream&controller=root

When it should just be:

  https://www.gov.uk/design-principles/mainstream

We also remove the option for supplying the text via anything other than a
block.  The app only ever calls this with a block, so there's no need to
support the other form.