Fix root controller test
We can't `get :styleguide` in the test because we don't define a route
for that action and rails 4 is stricter about that kind of thing than
rails 3.  We still have the controller action and views for this url,
but no route means they can't be rendered.  We'll work out later if
we should fully remove these remaining traces.