Disable sessions and stop exposing secrets we don't use
This app, like other frontend apps, does not use sessions so we don't need
to have a session store, and we don't need to configure valid values for
`secret_key_base` or `secret_token`.  However, rails will complain if we
don't set anything for `secret_key_base`, so we can follow the lead of
[frontend](https://github.com/alphagov/frontend/commit/98fb04c72473763087b02b2c5276eee13d8248df)
and set a random value from `SecureRandom.hex`.  We can remove the old
`secret_token` value too as we don't need it.