Deploy master branch builds to integration correctly
The previous version would send the wrong parameters, ending up with
`TARGET_APPLICATION` set to `-- Choose an app` instead of
`designprinciples`.  If we use the govuk.deployIntegration function
from the govuk jenkins library we avoid this problem.

We might want to look at replacing the entire jenkinsfile with a call
to `govuk.buildProject()`, but it's not clear if our existing script
wants all of that.